+++ src/app/tags/[tagName]/page.tsx
@@ -1,13 +1,13 @@
 import { Metadata } from 'next'
 import { notFound } from 'next/navigation'
 
-import { client } from '@/lib/hackersheet'
 
-import { DocumentList } from '@/components/document'
-import { Container } from '@/components/layout'
-import { PageHeader } from '@/components/layout'
-import { MutedMessage } from '@/components/message'
-import { ColorCircle, SmallTag } from '@/components/tag'
+import { MutedMessage } from '@/components/feedback/message'
+import { Container } from '@/components/layouts'
+import { PageHeader } from '@/components/layouts'
+import { DocumentList } from '@/features/document'
+import { ColorCircle, SmallTag } from '@/features/tag'
+import { client } from '@/lib/hackersheet'
 
 interface TagPageProps {
   params: Promise<{ tagName: string }>
