+++ src/components/controls/pagination/pagination.tsx
@@ -3,10 +3,10 @@
 import { useSearchParams } from 'next/navigation'
 import { useCallback, useEffect, useLayoutEffect, useRef, useState } from 'react'
 
+import { ScrollShadow } from '@/components/decorations/scroll-shadow'
 import { useClientOnly, useCreateQueryString } from '@/hooks'
 import { cn } from '@/lib/shadcn-utils'
 
-import { ScrollShadow } from '@/components/scroll-shadow'
 
 /**
  * ページネーション コンポーネントの Props
