+++ src/components/site/sns-list.tsx
@@ -1,8 +1,8 @@
 import { SiGithub, SiX, SiZenn } from 'react-icons/si'
 
+import { NextLink } from '@/components/navigations/link'
 import { Button } from '@/components/ui/button'
 
-import { NextLink } from '../link'
 
 /**
  * SnsList コンポーネント - 外部サービスへのリンクボタン群を表示します。
