+++ src/features/document/document-filter.tsx
@@ -1,8 +1,9 @@
 
-import SortBySelect from './sort-by-select'
-import { Input } from '../ui/input'
+import { SearchInput } from '@/components/forms/search-input'
+import { Input } from '@/components/ui/input'
 
-import { SearchInput } from '@/components/search-input'
+import SortBySelect from './sort-by-select'
+
 
 /**
  * DocumentFilter コンポーネントの Props
