+++ src/features/document/document-header.tsx
@@ -1,14 +1,14 @@
 import { CalendarIcon, ClockIcon, GitCommitHorizontalIcon, TagIcon } from 'lucide-react'
 
+import { FlowingGlow } from '@/components/decorations/effects'
+import { Link } from '@/components/navigations/link'
 import { HACKERSHEET_GITHUB_REPO_URL } from '@/constants'
+import { DocumentEmoji } from '@/features/document'
+import { SmallTag } from '@/features/tag'
 import { createDateFormat, timeAgo } from '@/utils'
 
 import type { Document } from '@/lib/hackersheet'
 
-import { DocumentEmoji } from '@/components/document'
-import { FlowingGlow } from '@/components/effects'
-import { Link } from '@/components/link'
-import { SmallTag } from '@/components/tag'
 
 
 /**
