+++ src/features/document/pickup-document-list.tsx
@@ -1,8 +1,8 @@
+import { MutedMessage } from '@/components/feedback/message'
 import { client } from '@/lib/hackersheet'
 
 import DocumentList from './document-list'
 
-import { MutedMessage } from '@/components/message'
 
 /**
  * PickupDocumentList コンポーネント - ピックアップされたドキュメントを表示します
