+++ src/tools/english-vocabulary/english-vocabulary.tsx
@@ -4,12 +4,12 @@ import { Check, Loader2, RotateCcw, X } from 'lucide-react'
 import { useCallback, useReducer, useState } from 'react'
 import { parse } from 'yaml'
 
+import { FlowingGlow } from '@/components/decorations/effects'
 import { Button } from '@/components/ui/button'
 import { cn } from '@/lib/shadcn-utils'
 
 import type { GameAction, GameState, QuestionSet } from './types'
 
-import { FlowingGlow } from '@/components/effects'
 
 /** 利用可能な問題集ファイル名 */
 const QUESTION_SET_FILES = [
